+++ lib/core/keys.dart
@@ -1,0 +1,4 @@
+class Keys {
+  static const String titleSplash = 'Welcome to Books App';
+  
+}
