--- lib/features/splash_screen/presentation/pages/splash_page.dart
@@ -1,16 +1,0 @@
-import 'package:flutter/material.dart';
-
-class SplashPage extends StatelessWidget {
-  const SplashPage({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return const Scaffold(
-      body: Center(
-        child: Text(
-          'Hello world',
-        ),
-      ),
-    );
-  }
-}
