+++ lib/main.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:library_app/features/splash_screen/presentation/pages/splash_page.dart';
+import 'package:library_app/features/splash/presentation/pages/splash_page.dart';
 
 void main() {
   runApp(
